1. refactor.

+++ src/main/java/fastdfs/client/FastdfsClient.java
@@ -362,7 +362,7 @@ public final class FastdfsClient implements Closeable {
      * @param size
      * @return
      */
-    public CompletableFuture<Void> download(String fileId, Object out, int offset, int size) {
+    public CompletableFuture<Void> download(String fileId, Object out, long offset, long size) {
         return download(FileId.fromString(fileId), out, offset, size);
     }
 
@@ -373,7 +373,7 @@ public final class FastdfsClient implements Closeable {
      * @param size
      * @return
      */
-    public CompletableFuture<Void> download(FileId fileId, Object out, int offset, int size) {
+    public CompletableFuture<Void> download(FileId fileId, Object out, long offset, long size) {
         Objects.requireNonNull(fileId, "fileId must not be null.");
         Objects.requireNonNull(out, "out must not be null.");
         return trackerClient
@@ -474,7 +474,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(String fileId, File file, int offset) {
+    public CompletableFuture<Void> modify(String fileId, File file, long offset) {
         return modify(FileId.fromString(fileId), file, offset);
     }
 
@@ -484,7 +484,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(FileId fileId, File file, int offset) {
+    public CompletableFuture<Void> modify(FileId fileId, File file, long offset) {
         Objects.requireNonNull(fileId, "fileId must not be null.");
         Objects.requireNonNull(file, "file must not be null.");
         return trackerClient
@@ -498,7 +498,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(String fileId, byte[] bytes, int offset) {
+    public CompletableFuture<Void> modify(String fileId, byte[] bytes, long offset) {
         return modify(FileId.fromString(fileId), bytes, offset);
     }
 
@@ -508,7 +508,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(FileId fileId, byte[] bytes, int offset) {
+    public CompletableFuture<Void> modify(FileId fileId, byte[] bytes, long offset) {
         return modify(fileId, bytes, bytes.length, offset);
     }
 
@@ -519,7 +519,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(String fileId, Object content, long size, int offset) {
+    public CompletableFuture<Void> modify(String fileId, Object content, long size, long offset) {
         return modify(FileId.fromString(fileId), content, size, offset);
     }
 
@@ -530,7 +530,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(FileId fileId, Object content, long size, int offset) {
+    public CompletableFuture<Void> modify(FileId fileId, Object content, long size, long offset) {
         Objects.requireNonNull(fileId, "fileId must not be null.");
         Objects.requireNonNull(content, "content must not be null.");
         return trackerClient
@@ -561,7 +561,7 @@ public final class FastdfsClient implements Closeable {
      * @param truncatedSize
      * @return
      */
-    public CompletableFuture<Void> truncate(String fileId, int truncatedSize) {
+    public CompletableFuture<Void> truncate(String fileId, long truncatedSize) {
         return truncate(FileId.fromString(fileId), truncatedSize);
     }
 
@@ -572,7 +572,7 @@ public final class FastdfsClient implements Closeable {
      * @param truncatedSize 截取字节数
      * @return
      */
-    public CompletableFuture<Void> truncate(FileId fileId, int truncatedSize) {
+    public CompletableFuture<Void> truncate(FileId fileId, long truncatedSize) {
         Objects.requireNonNull(fileId, "fileId must not be null.");
         return trackerClient
                 .updateStorageGet(fileId)
