+++ src/main/java/fastdfs/client/codec/FileTruncateEncoder.java
@@ -23,9 +23,9 @@ import static io.netty.util.CharsetUtil.UTF_8;
 public class FileTruncateEncoder implements Requestor.Encoder {
 
     private final FileId fileId;
-    private final int truncatedSize;
+    private final long truncatedSize;
 
-    public FileTruncateEncoder(FileId fileId, int truncatedSize) {
+    public FileTruncateEncoder(FileId fileId, long truncatedSize) {
         this.fileId = fileId;
         this.truncatedSize = truncatedSize;
     }
